Update to Kubernetes v1.23, add code for EKS EBS CSI driver, combine separate TAP cluster scripts into tap-cluster.sh

+++ tap-scripts/tanzu-essential-setup.sh
@@ -8,8 +8,6 @@ export TANZU_NET_API_TOKEN=$tanzu_net_api_token
 export INSTALL_REGISTRY_USERNAME=$tanzu_net_reg_user
 export INSTALL_REGISTRY_PASSWORD=$tanzu_net_reg_password
 
-
-
 export token=$(curl -X POST https://network.pivotal.io/api/v2/authentication/access_tokens -d '{"refresh_token":"'${TANZU_NET_API_TOKEN}'"}')
 access_token=$(echo ${token} | jq -r .access_token)
 
@@ -32,17 +30,20 @@ if [ "$os" == "$var" ]; then
 #file name - mac = tanzu-cluster-essentials-darwin-amd64-1.3.0.tgz , linux = tanzu-cluster-essentials-linux-amd64-1.3.0.tgz
 
 filename=tanzu-cluster-essentials-darwin-amd64-1.3.0.tgz
-tanzuessnurl=https://network.tanzu.vmware.com/api/v2/products/tanzu-cluster-essentials/releases/1180593/product_files/1330472/download
-mkdir $HOME/tanzu-cluster-essentials
-wget $tanzuessnurl --header="Authorization: Bearer ${access_token}" -O $HOME/tanzu-cluster-essentials/$filename
-tar -xvf $HOME/tanzu-cluster-essentials/$filename -C $HOME/tanzu-cluster-essentials
 
+if [[ -f "$HOME/tanzu-cluster-essentials/$filename" ]]; then
+    echo "$HOME/tanzu-cluster-essentials/$filename exists, skipping download"
+else
+    tanzuessnurl=https://network.tanzu.vmware.com/api/v2/products/tanzu-cluster-essentials/releases/1180593/product_files/1330472/download
+    mkdir $HOME/tanzu-cluster-essentials
+    wget $tanzuessnurl --header="Authorization: Bearer ${access_token}" -O $HOME/tanzu-cluster-essentials/$filename
+    tar -xvf $HOME/tanzu-cluster-essentials/$filename -C $HOME/tanzu-cluster-essentials
+
+    sudo cp $HOME/tanzu-cluster-essentials/kapp /usr/local/bin/kapp
+    sudo cp $HOME/tanzu-cluster-essentials/imgpkg /usr/local/bin/imgpkg
+fi
 cd $HOME/tanzu-cluster-essentials
 ./install.sh --yes
-
-sudo cp $HOME/tanzu-cluster-essentials/kapp /usr/local/bin/kapp
-sudo cp $HOME/tanzu-cluster-essentials/imgpkg /usr/local/bin/imgpkg
-   
 
 else
     echo "OS = Linux/ubuntu"
@@ -73,5 +74,3 @@ cd $HOME
 #chmod +x /usr/local/bin/demo-magic.sh
 
 #sudo apt install pv #required for demo-magic
-
-
